# openapi-3.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jiro Search API",
    "version": "0.2.0",
    "description": "Local-first web search, scraping, and social media intelligence platform. Supports 9+ search engines, 12 social platforms, hybrid search, structured extraction, and smart intent routing.",
    "contact": {
      "name": "Jiro Team",
      "url": "https://github.com/DevAnimecx/jiro"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.jiro.dev",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search the web",
        "description": "Search across multiple engines with filtering, biasing, and hybrid ranking.",
        "operationId": "search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/v1/search/multi": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Multi-query search",
        "description": "Execute multiple search queries in parallel.",
        "operationId": "searchMulti",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Multi-query results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiQueryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}